AII/backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.4.4
PyJWT==2.8.0
python-dotenv==1.0.0
PyMySQL==1.0.3
scikit-learn==1.7.2
xgboost==2.0.3
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
shap==0.44.1
requests==2.32.3
cryptography>=41.0.0
Flask-Mail
